requirements.txt
certifi==2016.2.28
flask=0.10.1
python-dateutil=2.5.3
requests==2.9.1
